pyproject.toml
[project]
name = "obligation-tracker"
version = "0.1.0"
description = "Contract Obligation Tracker & Compliance Dashboard"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Noam Raz"},
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "anthropic>=0.42.0",
    "python-docx>=1.1.0",
    "pymupdf>=1.25.0",
    "python-multipart>=0.0.20",
    "streamlit>=1.41.0",
    "streamlit-calendar>=1.2.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["ot_app*", "ot_frontend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
